update homework week3

+++ Week3/homework/ex2/transactions-create-tables.js
@@ -19,7 +19,7 @@ con.connect(() => {
   con.query(`CREATE TABLE account_changes (
     change_number INT PRIMARY KEY AUTO_INCREMENT, 
     account_number INT,
-    amount INT,
+    amount DECIMAL,
     changed_date DATE,
     remark VARCHAR(128),
     FOREIGN KEY (account_number) REFERENCES account(account_number)
